Add pinning tests for grammar
Fixes #149

- Converts `test.ts` into a set of pinning tests that can be run using `npm test`
- Adds a travis file so that we can setup these tests to be run on CI
