+++ .vscodeignore
@@ -1,2 +1,3 @@
 .vscode/**
 .gitignore
+test/
