+++ test/colorize-fixtures/arrow-function.js
@@ -1,0 +1,3 @@
+const arrowFun = (...args) => css`
+  height: 12px;
+`;
