+++ test/colorize-fixtures/attrs.js
@@ -1,0 +1,5 @@
+const Link = styled.a.attrs({
+    target: "_blank"
+})`
+  color: red;
+`;
