+++ test/colorize-fixtures/comment.js
@@ -1,0 +1,4 @@
+// const comment = css`
+//     height: 20px;
+//     padding: 5px;
+// `
