+++ test/colorize-fixtures/component.js
@@ -1,0 +1,3 @@
+const Component = styled(Component)`
+  color: #ebebeb;
+`;
