+++ test/colorize-fixtures/createGlobalStyle.js
@@ -1,0 +1,5 @@
+const GlobalStyles = createGlobalStyle`
+  html {
+    color: 'red';
+  }
+`;
