+++ test/colorize-fixtures/dot-tag.js
@@ -1,0 +1,6 @@
+const DotTag = styled.div`
+  color: #ebebeb;
+  font-size: ${props => props.theme.fontSize};
+  font-family: Helvetica;
+  ${mixin};
+`;
