+++ test/colorize-fixtures/export-default.js
@@ -1,0 +1,3 @@
+export default styled(ExportComponent)`
+  max-width: 100%;
+`;
