+++ test/colorize-fixtures/extend.js
@@ -1,0 +1,7 @@
+const Comp = styled.div`
+  color: red;
+`;
+
+const NewComp = Comp.extend`
+  color: green;
+`;
