+++ test/colorize-fixtures/injectGlobal.js
@@ -1,0 +1,10 @@
+injectGlobal`
+  @font-face {
+    font-family: "Operator Mono";
+    src: url("../fonts/Operator-Mono.ttf");
+  }
+
+  body {
+    margin: 0;
+  }
+`;
