+++ test/colorize-fixtures/inside-function.js
@@ -1,0 +1,5 @@
+function insideFunction() {
+    return styled.div`
+    height: 15px;
+  `;
+}
