+++ test/colorize-fixtures/inside-method.js
@@ -1,0 +1,7 @@
+class InsideMethod extends React.Component {
+    render() {
+        return styled(Component)`
+      line-height: 21px;
+    `;
+    }
+}
