+++ test/colorize-fixtures/keyframes.js
@@ -1,0 +1,8 @@
+const rotate360 = keyframes`
+    from {
+        transform: rotate(0deg);
+    }
+    to {
+        transform: rotate(1turn);
+    }
+`;
