+++ test/colorize-fixtures/mixin.js
@@ -1,0 +1,4 @@
+const mixin = css`
+  height: 20px;
+  padding: 5px;
+`;
