+++ test/colorize-fixtures/segmented-component.js
@@ -1,0 +1,3 @@
+const SegmentedComponent = styled(Segmented.Component)`
+  padding: 3px;
+`;
