+++ test/colorize-fixtures/string-tagname.js
@@ -1,0 +1,3 @@
+const StringTagname = styled("div")`
+  color: #ff0000;
+`;
