+++ test/colorize-fixtures/typescript-emotion.ts
@@ -1,0 +1,6 @@
+// Typescript, Emotion
+// prettier-ignore
+const Container = styled<ContainerProps, 'div'>('div')`
+  height: 50px;
+  display: ${(props) => props.display};
+`
