+++ test/colorize-fixtures/typescript.ts
@@ -1,0 +1,3 @@
+const Root = styled<RootProps>("div")`
+  height: ${props => (props.label ? 72 : 48)}px;
+`;
