+++ test/colorize-fixtures/variable-assignment.js
@@ -1,0 +1,10 @@
+let variableAssignment;
+
+variableAssignment = css`
+  height: 1px;
+`(
+    /* expression */
+    styled.div`
+    height: 12px;
+  `
+);
