+++ test/colorize-fixtures/withComponent.js
@@ -1,0 +1,15 @@
+const NewCompWithString = CompWithComponent.withComponent("span").extend`
+  color: green;
+`;
+
+const NewCompWithString = CompWithComponent.withComponent("span")`
+  color: green;
+`;
+
+const NewCompWithStringOneLine = CompWithComponent.withComponent(
+    "span"
+)`color: green;`;
+
+const NewCompWithComponent = CompWithComponent.withComponent(OtherComp)`
+  color: green;
+`;
